+++ plugins/geometry_vtk_streamlines.cpp
@@ -25,25 +25,61 @@
 #include "plugin.h"
 
 #ifdef PLUGIN_VTK_STREAMLINES // XXX
-#include <vtkPolyData.h>
-#include <vtkDataSetReader.h>
 #include <vtkSmartPointer.h>
+#include <vtkPolyData.h>
+#include <vtkPointData.h>
+#include <vtkDataSetReader.h>
+#include <vtkColorTransferFunction.h>
 #endif
 
 extern "C"
 void
 load_file(GenerateFunctionResult &result, PluginState *state)
 {
-    const json& parameters = state->parameters;
+    //char        msg[1024];
 
+    const json& parameters = state->parameters;
     const std::string& file = parameters["file"];
-    char        msg[1024];
+
+    float radius = 1.0f;
+    if (parameters.find("radius") != parameters.end())
+        radius = parameters["radius"];    
+
+    vtkSmartPointer<vtkColorTransferFunction> cool2warm;
+    bool        color_by_scalars = false;
+    std::string scalars;
+    float       scalar_range[2];
+    float       scalar_value_range;
+
+    scalar_range[0] = 0.0f;
+    scalar_range[1] = 1.0f;
+    if (parameters.find("scalar_range") != parameters.end())
+    {
+        auto& p_scalar_range = parameters["scalar_range"];
+        scalar_range[0] = p_scalar_range[0];
+        scalar_range[1] = p_scalar_range[1];        
+    }
+    scalar_value_range = scalar_range[1] - scalar_range[0];
+    
+    if (parameters.find("scalars") != parameters.end())
+    {
+        scalars = parameters["scalars"];        
+
+        cool2warm = vtkSmartPointer<vtkColorTransferFunction>::New();
+        cool2warm->SetColorSpaceToDiverging();
+        cool2warm->AddRGBPoint(0.0, 0.230, 0.299, 0.754);
+        cool2warm->AddRGBPoint(1.0, 0.706, 0.016, 0.150);
+
+        color_by_scalars = true;
+        printf("... Using scalar data range %.6f, %.6f\n", scalar_range[0], scalar_range[1]);
+    }
 
     printf("... Loading %s\n", file.c_str());
 
     vtkSmartPointer<vtkDataSetReader> reader = vtkSmartPointer<vtkDataSetReader>::New();
     reader->SetFileName(file.c_str());
     reader->Update();
+    // XXX check file was read correctly
 
     /*
     if (!scene)
@@ -55,7 +91,20 @@ load_file(GenerateFunctionResult &result, PluginState *state)
     }
     */
 
-    vtkSmartPointer<vtkDataSet> dataset = reader->GetOutput();
+    vtkSmartPointer<vtkDataSet> dataset = reader->GetOutput();    
+    vtkSmartPointer<vtkDataArray> scalar_data;
+
+    if (color_by_scalars)
+    {
+        vtkSmartPointer<vtkPointData> point_data = dataset->GetPointData();
+        point_data->SetActiveScalars(scalars.c_str());
+        scalar_data = point_data->GetScalars();
+        if (scalar_data == nullptr)
+        {
+            printf("... WARNING: could not find point data scalar array '%s'!\n", scalars.c_str());
+            color_by_scalars = false;
+        }
+    }
 
     const vtkIdType np = dataset->GetNumberOfPoints();
     const vtkIdType nc = dataset->GetNumberOfCells();
@@ -65,8 +114,10 @@ load_file(GenerateFunctionResult &result, PluginState *state)
 
     std::vector<float>      positions;
     std::vector<uint32_t>   indices;
+    std::vector<float>      vcolors;
     uint32_t    curindex = 0;
-    double  p[3];
+    double  p[3], *col;
+    float v, f;
 
     for (vtkIdType i = 0; i < nc; i++)
     {
@@ -91,7 +142,7 @@ load_file(GenerateFunctionResult &result, PluginState *state)
         positions.push_back(p[0]);
         positions.push_back(p[1]);
         positions.push_back(p[2]);
-        
+
         dataset->GetPoint(qi, p);
         positions.push_back(p[0]);
         positions.push_back(p[1]);
@@ -100,14 +151,35 @@ load_file(GenerateFunctionResult &result, PluginState *state)
         indices.push_back(curindex);
         
         curindex += 2;
+
+        if (color_by_scalars)
+        {
+            v = (float)(scalar_data->GetTuple1(pi));
+            f = (v - scalar_range[0]) / scalar_value_range;
+            col = cool2warm->GetColor(f);
+            vcolors.push_back(col[0]);
+            vcolors.push_back(col[1]);
+            vcolors.push_back(col[2]);
+            vcolors.push_back(f);           // Note: alpha = normalized value
+
+            v = (float)(scalar_data->GetTuple1(qi));
+            f = (v - scalar_range[0]) / scalar_value_range;
+            col = cool2warm->GetColor(f);
+            vcolors.push_back(col[0]);
+            vcolors.push_back(col[1]);
+            vcolors.push_back(col[2]);
+            vcolors.push_back(f);
+        }
     }
+
+    OSPData data;
 
     OSPGeometry geometry = ospNewGeometry("streamlines");
     {
         printf("... %d positions\n", positions.size()/3);
         printf("... %d indices\n", indices.size());
 
-        OSPData data = ospNewCopiedData(positions.size()/3, OSP_VEC3F, positions.data());
+        data = ospNewCopiedData(positions.size()/3, OSP_VEC3F, positions.data());
         ospCommit(data);
         ospSetObject(geometry, "vertex.position", data);
         ospRelease(data);
@@ -118,137 +190,16 @@ load_file(GenerateFunctionResult &result, PluginState *state)
         ospRelease(data);
     }
 
-    float radius = 1.0f;
-    if (parameters.find("radius") != parameters.end())
-        radius = parameters["radius"];
-
     ospSetFloat(geometry, "radius", radius);
-
     ospSetBool(geometry, "smooth", false);
 
-#if 0
-    const int&  nvertices = mesh->mNumVertices;
-    float       min[3] = {1e6, 1e6, 1e6}, max[3] = {-1e6, -1e6, -1e6};
-    float       x, y, z;
-
-    std::vector<float> vertices;
-    std::vector<uint32_t> triangles;
-
-    // Triangle mesh
-    OSPGeometry geometry = ospNewGeometry("triangles");
+    if (color_by_scalars)
     {
-        printf("... %d vertices\n", nvertices);
-
-        // Vertices
-        auto& mv = mesh->mVertices;
-
-        for (int i = 0; i < nvertices; i++)
-        {
-            x = mv[i].x;
-            y = mv[i].y;
-            z = mv[i].z;
-
-            vertices.push_back(x);
-            vertices.push_back(y);
-            vertices.push_back(z);
-
-            min[0] = std::min(min[0], x);
-            min[1] = std::min(min[1], y);
-            min[2] = std::min(min[2], z);
-
-            max[0] = std::max(max[0], x);
-            max[1] = std::max(max[1], y);
-            max[2] = std::max(max[2], z);
-        }
-
-        OSPData data = ospNewCopiedData(nvertices, OSP_VEC3F, vertices.data());
-        ospCommit(data);
-        ospSetObject(geometry, "vertex.position", data);
-        //ospRelease(data);
-    }
-
-    if (mesh->HasFaces())
-    {
-        printf("... %d triangles\n",  mesh->mNumFaces);
-
-        // Triangles
-        auto& ff = mesh->mFaces;
-
-        for (int i = 0; i < mesh->mNumFaces; i++)
-        {
-            triangles.push_back(ff[i].mIndices[0]);
-            triangles.push_back(ff[i].mIndices[1]);
-            triangles.push_back(ff[i].mIndices[2]);
-        }
-
-        // XXX check /3
-        OSPData data = ospNewCopiedData(triangles.size()/3, OSP_VEC3UI, triangles.data());
-        ospCommit(data);
-        ospSetObject(geometry, "index", data);
-        //ospRelease(data);
-    }
-    else
-        printf("... Mesh has no faces?\n");
-
-    if (mesh->HasVertexColors(0))
-    {
-        printf("... Mesh has vertex colors\n");
-
-        std::vector<float>  colors;
-        auto& vc = mesh->mColors[0];
-
-        for (int i = 0; i < nvertices; i++)
-        {
-            colors.push_back(vc[i].r);
-            colors.push_back(vc[i].g);
-            colors.push_back(vc[i].b);
-            colors.push_back(vc[i].a);
-        }
-
-        OSPData data = ospNewCopiedData(nvertices, OSP_VEC4F, colors.data());
+        data = ospNewCopiedData(vcolors.size()/4, OSP_VEC4F, vcolors.data());
         ospCommit(data);
         ospSetObject(geometry, "vertex.color", data);
+        ospRelease(data);
     }
-
-    if (mesh->HasNormals())
-    {
-        printf("... Mesh has normals\n");
-
-        std::vector<float>  normals;
-        auto& nn = mesh->mNormals;
-
-        for (int i = 0; i < nvertices; i++)
-        {
-            normals.push_back(nn[i].x);
-            normals.push_back(nn[i].y);
-            normals.push_back(nn[i].z);
-        }
-
-        OSPData data = ospNewCopiedData(nvertices, OSP_VEC3F, normals.data());
-        ospCommit(data);
-        ospSetObject(geometry, "vertex.normal", data);
-    }
-
-    if (mesh->HasTextureCoords(0))
-    {
-        printf("... Mesh has texture coordinates\n");
-
-        std::vector<float>  texcoords;
-        auto& tc = mesh->mTextureCoords[0];
-
-        for (int i = 0; i < nvertices; i++)
-        {
-            texcoords.push_back(tc[i].x);
-            texcoords.push_back(tc[i].y);
-            texcoords.push_back(tc[i].z);
-        }
-
-        OSPData data = ospNewCopiedData(nvertices, OSP_VEC2F, texcoords.data());
-        ospCommit(data);
-        ospSetObject(geometry, "vertex.texcoord", data); 
-    }
-    
-#endif    
 
     ospCommit(geometry);
 
@@ -267,8 +218,9 @@ static PluginParameters
 parameters = {
     
     {"file",        PARAM_STRING,   1, FLAG_NONE, "VTK file to load"},
-    {"radius",      PARAM_FLOAT,    1, FLAG_NONE, "Radius"},
-    //{"divisions",     PARAM_INT,        1, FLAG_NONE, "Divisions (for QC simplification)"},
+    {"radius",      PARAM_FLOAT,    1, FLAG_NONE, "Radius"},    
+    {"scalars",     PARAM_STRING,   1, FLAG_OPTIONAL, "Scalar values to show (array name)"},
+    {"scalar_range", PARAM_FLOAT,   2, FLAG_OPTIONAL, "Scalar value range to use for coloring"},
         
     PARAMETERS_DONE         // Sentinel (signals end of list)
 };